Dirty fixed an updater bug
Node.js reads the registry value as %userprofile% literally, saves updater files to the local tribeslauncher directory/%userprofile%/*updatedfiles*.
os.homedir() grabs the user profile directory, this probably should be changed into a string that can be edited by the config to allow users to fix on their own.
This might not work if the documents folder is moved to a location that is within the %userprofile% but isn't called Documents or moved to some random folder. People aren't silly enough to do that, right, right?